1.2

+++ barinove.egor/T3/Polygon.cpp
@@ -13,7 +13,7 @@ double area(const Polygon& p) {
     }
 
     double sum = 0;
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         int j = (i + 1) % n;
         sum += pts[i].x * pts[j].y;
         sum -= pts[i].y * pts[j].x;
@@ -72,3 +72,4 @@ std::vector<Polygon> readShapesFromFile(const std::string& filename) {
 
     return result;
 }
+
